# Conflicts:
#	TelegramUI.xcodeproj/project.pbxproj
#	TelegramUI/ChatController.swift
#	TelegramUI/ChatControllerNode.swift
#	TelegramUI/ChatMessageInteractiveMediaNode.swift
#	TelegramUI/OpenResolvedUrl.swift
#	TelegramUI/PeerMediaCollectionControllerNode.swift
#	TelegramUI/PresentationData.swift
#	TelegramUI/PresentationStrings.swift
#	TelegramUI/PresentationThemeSettings.swift
#	TelegramUI/Resources/PresentationStrings.mapping
#	TelegramUI/SettingsThemeWallpaperNode.swift
#	TelegramUI/TelegramApplicationContext.swift
#	TelegramUI/TelegramInitializeLegacyComponents.swift
#	TelegramUI/TelegramRootController.swift
#	TelegramUI/ThemeColorsGridController.swift
#	TelegramUI/ThemeGridController.swift
#	TelegramUI/ThemeGridControllerItem.swift
#	TelegramUI/ThemeGridControllerNode.swift
#	TelegramUI/ThemeGridSearchContentNode.swift
#	TelegramUI/ThemeSettingsChatPreviewItem.swift
#	TelegramUI/ThemeSettingsController.swift
#	TelegramUI/WallpaperGalleryController.swift
#	TelegramUI/WallpaperGalleryItem.swift
#	TelegramUI/WallpaperListPreviewController.swift
#	TelegramUI/WallpaperListPreviewControllerNode.swift

+++ TelegramUI/ThemeGridSelectionPanelNode.swift
@@ -13,7 +13,7 @@ final class ThemeGridSelectionPanelNode: ASDisplayNode {
     
     private var theme: PresentationTheme
     
-    var selectedIndices = Set<Int>() {
+    var selectedIndices = Set<Int64>() {
         didSet {
             if oldValue != self.selectedIndices {
                 self.deleteButton.isEnabled = !self.selectedIndices.isEmpty
